Fixed bug. Double converted to int, so zoom was 0 -> black bmp

+++ Fractal Image Drawer/Zoom.h
@@ -6,6 +6,6 @@ struct Zoom
 	int _y{ 0 };
 	double _scale{ 0.0 };
 
-	Zoom(int x, int y, int scale) : _x(x), _y(y), _scale(scale) {}
+	Zoom(int x, int y, double scale) : _x(x), _y(y), _scale(scale) {}
 };
